Move web kernel to web namespace
