+++ .env.test
@@ -1,5 +1,5 @@
 # define your env variables for the test env here
-KERNEL_CLASS='App\Kernel'
+KERNEL_CLASS='VDOLog\Web\Kernel'
 APP_SECRET='s$cretf0rt3st'
 APP_LOCALE=de
 SYMFONY_DEPRECATIONS_HELPER=999999
