+++ behat.yml.dist
@@ -2,13 +2,15 @@ default:
     suites:
         default:
             contexts:
-                - App\Tests\Specifications\GameContext
-                - App\Tests\Specifications\ProtocolContext
+                - VDOLog\Tests\Specifications\GameContext
+                - VDOLog\Tests\Specifications\ProtocolContext
                 - Behat\MinkExtension\Context\MinkContext
-                - App\Tests\Specifications\FormContext
+                - VDOLog\Tests\Specifications\FormContext
 
     extensions:
-        FriendsOfBehat\SymfonyExtension: null
+        FriendsOfBehat\SymfonyExtension:
+            kernel:
+                class: VDOLog\Web\Kernel
         Behat\MinkExtension:
             sessions:
                 symfony:
