+++ config/packages/doctrine.yaml
@@ -11,6 +11,6 @@ doctrine:
             App:
                 is_bundle: false
                 type: annotation
-                dir: '%kernel.project_dir%/src/Entity'
-                prefix: 'App\Entity'
+                dir: '%kernel.project_dir%/src/Web/Entity'
+                prefix: 'VDOLog\Web\Entity'
                 alias: App
