+++ config/routes/annotations.yaml
@@ -1,3 +1,3 @@
 controllers:
-    resource: ../../src/Controller/
+    resource: ../../src/Web/Controller/
     type: annotation
