+++ src/Web/Controller/DashboardController.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace VDOLog\Web\Controller;
 
-use App\Entity\Game;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use VDOLog\Web\Entity\Game;
 
 class DashboardController extends AbstractController
 {
