+++ src/Web/Form/GameType.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace App\Form;
+namespace VDOLog\Web\Form;
 
-use App\Entity\Game;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints\NotBlank;
+use VDOLog\Web\Entity\Game;
 
 class GameType extends AbstractType
 {
