+++ src/Web/Form/ProtocolType.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace App\Form;
+namespace VDOLog\Web\Form;
 
-use App\Entity\Protocol;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\CallbackTransformer;
@@ -14,6 +13,7 @@ use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints\NotBlank;
+use VDOLog\Web\Entity\Protocol;
 
 final class ProtocolType extends AbstractType
 {
