+++ src/Web/Kernel.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App;
+namespace VDOLog\Web;
 
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
 use Symfony\Component\Config\Loader\LoaderInterface;
