+++ src/Web/Model/ApplicationConfiguration.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Model;
+namespace VDOLog\Web\Model;
 
 use Assert\Assertion;
